requirements.txt
faiss
fastapi
insightface
numpy
Pillow
python-dotenv
uvicorn
onnxruntime